+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Head
+* Added: support for regular expression property identification in `property-blacklist`.
 * Added: `codeFilename` option to Node API.
 * Added: `font-weight-notation` rule.
 * Added: `media-feature-no-missing-punctuation` rule.